+++ library/Episciences/PapersManager.php
@@ -3964,7 +3964,7 @@ class Episciences_PapersManager
         try {
             $db = Zend_Db_Table_Abstract::getDefaultAdapter();
             $paper = self::get($docId, false);
-            $toJson = $paper->toJson(Episciences_Paper_XmlExportManager::ALL_KEY);
+            $toJson = $paper->toJson();
             $str = sprintf('UPDATE `PAPERS` set `DOCUMENT` = %s  WHERE DOCID = %s;', $db->quote($toJson), $docId);
             $db->query($str)->closeCursor();
         } catch (Zend_Db_Statement_Exception $e) {